Renaming of classes Refractor
