+++ src/BorrowBookControl.java
@@ -3,16 +3,16 @@ import java.util.List;
 
 public class BorrowBookControl {
 	
-	private BorrowBookUI ui;
+	private BorrowBookUi ui;
 	
-	private library L;
-	private member M;
+	private Library L;
+	private Member M;
 	private enum CONTROL_STATE { INITIALISED, READY, RESTRICTED, SCANNING, IDENTIFIED, FINALISING, COMPLETED, CANCELLED };
 	private CONTROL_STATE state;
 	
-	private List<book> PENDING;
-	private List<loan> COMPLETED;
-	private book B;
+	private List<Book> PENDING;
+	private List<Loan> COMPLETED;
+	private Book B;
 	
 	
 	public BorrowBookControl() {
@@ -21,12 +21,12 @@ public class BorrowBookControl {
 	}
 	
 
-	public void setUI(BorrowBookUI ui) {
+	public void setUI(BorrowBookUi ui) {
 		if (!state.equals(CONTROL_STATE.INITIALISED)) 
 			throw new RuntimeException("BorrowBookControl: cannot call setUI except in INITIALISED state");
 			
 		this.ui = ui;
-		ui.setState(BorrowBookUI.UI_STATE.READY);
+		ui.setState(BorrowBookUi.UI_STATE.READY);
 		state = CONTROL_STATE.READY;		
 	}
 
@@ -42,12 +42,12 @@ public class BorrowBookControl {
 		}
 		if (L.memberCanBorrow(M)) {
 			PENDING = new ArrayList<>();
-			ui.setState(BorrowBookUI.UI_STATE.SCANNING);
+			ui.setState(BorrowBookUi.UI_STATE.SCANNING);
 			state = CONTROL_STATE.SCANNING; }
 		else 
 		{
 			ui.display("Member cannot borrow at this time");
-			ui.setState(BorrowBookUI.UI_STATE.RESTRICTED); }}
+			ui.setState(BorrowBookUi.UI_STATE.RESTRICTED); }}
 	
 	
 	public void Scanned(int bookId) {
@@ -65,7 +65,7 @@ public class BorrowBookControl {
 			return;
 		}
 		PENDING.add(B);
-		for (book B : PENDING) {
+		for (Book B : PENDING) {
 			ui.display(B.toString());
 		}
 		if (L.loansRemainingForMember(M) - PENDING.size() == 0) {
@@ -81,11 +81,11 @@ public class BorrowBookControl {
 		}
 		else {
 			ui.display("\nFinal Borrowing List");
-			for (book b : PENDING) {
+			for (Book b : PENDING) {
 				ui.display(b.toString());
 			}
-			COMPLETED = new ArrayList<loan>();
-			ui.setState(BorrowBookUI.UI_STATE.FINALISING);
+			COMPLETED = new ArrayList<Loan>();
+			ui.setState(BorrowBookUi.UI_STATE.FINALISING);
 			state = CONTROL_STATE.FINALISING;
 		}
 	}
@@ -95,21 +95,21 @@ public class BorrowBookControl {
 		if (!state.equals(CONTROL_STATE.FINALISING)) {
 			throw new RuntimeException("BorrowBookControl: cannot call commitLoans except in FINALISING state");
 		}	
-		for (book b : PENDING) {
-			loan loan = L.issueLoan(b, M);
-			COMPLETED.add(loan);			
+		for (Book b : PENDING) {
+			Loan Loan = L.issueLoan(b, M);
+			COMPLETED.add(Loan);
 		}
 		ui.display("Completed Loan Slip");
-		for (loan loan : COMPLETED) {
-			ui.display(loan.toString());
+		for (Loan Loan : COMPLETED) {
+			ui.display(Loan.toString());
 		}
-		ui.setState(BorrowBookUI.UI_STATE.COMPLETED);
+		ui.setState(BorrowBookUi.UI_STATE.COMPLETED);
 		state = CONTROL_STATE.COMPLETED;
 	}
 
 	
 	public void cancel() {
-		ui.setState(BorrowBookUI.UI_STATE.CANCELLED);
+		ui.setState(BorrowBookUi.UI_STATE.CANCELLED);
 		state = CONTROL_STATE.CANCELLED;
 	}
