+++ src/Loan.java
@@ -1,0 +1,78 @@
+import java.io.Serializable;
+import java.text.SimpleDateFormat;
+import java.util.Date;
+
+@SuppressWarnings("serial")
+public class Loan implements Serializable {
+	
+	public static enum LOAN_STATE { CURRENT, OVER_DUE, DISCHARGED };
+	
+	private int ID;
+	private Book B;
+	private Member M;
+	private Date D;
+	private LOAN_STATE state;
+
+	
+	public Loan(int loanId, Book book, Member Member, Date dueDate) {
+		this.ID = loanId;
+		this.B = book;
+		this.M = Member;
+		this.D = dueDate;
+		this.state = LOAN_STATE.CURRENT;
+	}
+
+	
+	public void checkOverDue() {
+		if (state == LOAN_STATE.CURRENT &&
+			Calendar.getInstance().Date().after(D)) {
+			this.state = LOAN_STATE.OVER_DUE;			
+		}
+	}
+
+	
+	public boolean isOverDue() {
+		return state == LOAN_STATE.OVER_DUE;
+	}
+
+	
+	public Integer getId() {
+		return ID;
+	}
+
+
+	public Date getDueDate() {
+		return D;
+	}
+	
+	
+	public String toString() {
+		SimpleDateFormat sdf = new SimpleDateFormat("dd/MM/yyyy");
+
+		StringBuilder sb = new StringBuilder();
+		sb.append("Loan:  ").append(ID).append("\n")
+		  .append("  Borrower ").append(M.getId()).append(" : ")
+		  .append(M.getLastName()).append(", ").append(M.getFirstName()).append("\n")
+		  .append("  Book ").append(B.ID()).append(" : " )
+		  .append(B.Title()).append("\n")
+		  .append("  DueDate: ").append(sdf.format(D)).append("\n")
+		  .append("  State: ").append(state);		
+		return sb.toString();
+	}
+
+
+	public Member Member() {
+		return M;
+	}
+
+
+	public Book Book() {
+		return B;
+	}
+
+
+	public void Loan() {
+		state = LOAN_STATE.DISCHARGED;		
+	}
+
+}
