+++ src/PayFineControl.java
@@ -1,11 +1,11 @@
 public class PayFineControl {
 	
-	private PayFineUI ui;
+	private PayFineUi ui;
 	private enum CONTROL_STATE { INITIALISED, READY, PAYING, COMPLETED, CANCELLED };
 	private CONTROL_STATE state;
 	
-	private library library;
-	private member member;;
+	private Library library;
+	private Member Member;;
 
 
 	public PayFineControl() {
@@ -14,12 +14,12 @@ public class PayFineControl {
 	}
 	
 	
-	public void setUI(PayFineUI ui) {
+	public void setUI(PayFineUi ui) {
 		if (!state.equals(CONTROL_STATE.INITIALISED)) {
 			throw new RuntimeException("PayFineControl: cannot call setUI except in INITIALISED state");
 		}	
 		this.ui = ui;
-		ui.setState(PayFineUI.UI_STATE.READY);
+		ui.setState(PayFineUi.UI_STATE.READY);
 		state = CONTROL_STATE.READY;		
 	}
 
@@ -28,20 +28,20 @@ public class PayFineControl {
 		if (!state.equals(CONTROL_STATE.READY)) {
 			throw new RuntimeException("PayFineControl: cannot call cardSwiped except in READY state");
 		}	
-		member = library.getMember(memberId);
+		Member = library.getMember(memberId);
 		
-		if (member == null) {
+		if (Member == null) {
 			ui.display("Invalid Member Id");
 			return;
 		}
-		ui.display(member.toString());
-		ui.setState(PayFineUI.UI_STATE.PAYING);
+		ui.display(Member.toString());
+		ui.setState(PayFineUi.UI_STATE.PAYING);
 		state = CONTROL_STATE.PAYING;
 	}
 	
 	
 	public void cancel() {
-		ui.setState(PayFineUI.UI_STATE.CANCELLED);
+		ui.setState(PayFineUi.UI_STATE.CANCELLED);
 		state = CONTROL_STATE.CANCELLED;
 	}
 
@@ -50,12 +50,12 @@ public class PayFineControl {
 		if (!state.equals(CONTROL_STATE.PAYING)) {
 			throw new RuntimeException("PayFineControl: cannot call payFine except in PAYING state");
 		}	
-		double change = member.payFine(amount);
+		double change = Member.payFine(amount);
 		if (change > 0) {
 			ui.display(String.format("Change: $%.2f", change));
 		}
-		ui.display(member.toString());
-		ui.setState(PayFineUI.UI_STATE.COMPLETED);
+		ui.display(Member.toString());
+		ui.setState(PayFineUi.UI_STATE.COMPLETED);
 		state = CONTROL_STATE.COMPLETED;
 		return change;
 	}
